+++ CHANGELOG.md
@@ -36,7 +36,7 @@
 
 - Modules [8FACE, 8FACEx2](./docs/EightFace.md)
     - Added "Auto"-mode besides "Read" and "Write" ([manual](./docs/EightFace.md#auto-mode)) (#251)
-    - Added "Shift back" context menu option (#275)
+    - Added "Shift front" and "Shift back" context menu options (#275)
 - Module [INTERMIX](./docs/Intermix.md)
     - Added context menu option "Scene lock" to prevent accidental changes
 - Module [MIDI-CAT](./docs/MidiCat.md)